requirements.txt
flask==1.1.2
flask-marshmallow==0.14.0
flask-sqlalchemy==2.5.0
marshmallow-sqlalchemy==0.24.1

